Use global AppLogger instead of console.log
This commit centralizes the logging mechanism instead of using console.log/error/.... This way, we are now logging to a file when using the repl to avoid having random logs pop out while using the REPL. When only using the CLI, we're still using the console. We can now also edit all the console logging in a single file rather than modifying all the console calls.